# rt: optimize spawn functions