--- a/tokio/src/runtime/handle.rs
+++ b/tokio/src/runtime/handle.rs
@@ -154,6 +154,7 @@ impl Handle {
     /// # }
     /// ```
     #[track_caller]
+    #[inline]
     pub fn spawn<F>(&self, future: F) -> JoinHandle<F::Output>
     where
         F: Future + Send + 'static,