--- a/tokio/src/runtime/task/raw.rs
+++ b/tokio/src/runtime/task/raw.rs
@@ -1,5 +1,5 @@
 use crate::future::Future;
-use crate::runtime::task::{Cell, Harness, Header, Schedule, State};
+use crate::runtime::task::{Harness, Header, Schedule};
 
 use std::ptr::NonNull;
 use std::task::{Poll, Waker};
@@ -48,17 +48,6 @@ pub(super) fn vtable<T: Future, S: Schedule>() -> &'static Vtable {
 }
 
 impl RawTask {
-    pub(super) fn new<T, S>(task: T, scheduler: S) -> RawTask
-    where
-        T: Future,
-        S: Schedule,
-    {
-        let ptr = Box::into_raw(Cell::<_, S>::new(task, scheduler, State::new()));
-        let ptr = unsafe { NonNull::new_unchecked(ptr as *mut Header) };
-
-        RawTask { ptr }
-    }
-
     pub(super) unsafe fn from_raw(ptr: NonNull<Header>) -> RawTask {
         RawTask { ptr }
     }