@yiyuanliu 2022-02-11
> I'm not a big fan of the `UninitTask` strategy. How about the following alternative:
> 
> 1. Create an "owning" raw-pointer type which can own a value on the stack, but be pointer-sized itself.
> 2. Instead of passing around the future, pass around the owning stack pointer.
> 3. After allocating memory, move ownership of the future from the stack into the allocation.

I have considered this option before finally moving to `UninitTask`.

I think `UninitTask` gives the compiler more opportunity to do optimization. For owned stack pointers, there are still many function calls before the future is moved to the heap. The optimization compiler can do are limited. I'm not sure but I guess in this case the future will still be initialized on stack and then copied to heap. Although performance is improved compared to the existing code, there is still a memory copy.

With `UninitTask`, the compiler can sometimes even initialize futures on heap directly. While this optimization doesn't always work, in benchmark it appears there is still some data copy as spawn performance drops as the future size grows, but `UninitTask` does give the compiler more opportunity to optimize.