@Darksonn 2022-10-12
The PR probably is not getting merged in its current state. It adds a lot of complexity, and I still believe in the alternate strategy that I suggested earlier. Rust isn't even able to elide the init-on-stack-first with `Box::new(LargeStruct { ... })`, so I don't see how it would ever be able to do so in `tokio::spawn`.